+++ shadow_mapping/main.cpp
@@ -259,11 +259,11 @@ void Win::handle_key_press(std::string key, bool pressed) {
 }
 
 void Win::handle_mouse_motion(int x, int y) {
-	float dt = delta_time;
+	float dt = (float)delta_time;
 	if(dt < 0.01)
-		dt = 0.01;
+		dt = 0.01f;
 	if(dt > 1.0)
-		dt = 1.0;
+		dt = 1.0f;
 	if (rel_mode) {
 		camera.yaw(-glm::atan((float)x) * dt);
 		camera.pitch(-glm::atan((float)y) * dt);
